I2c inteface now uses byte instead of uint16

+++ platforms/i2c/i2c.go
@@ -2,6 +2,6 @@ package i2c
 
 type I2cInterface interface {
 	I2cStart(byte)
-	I2cRead(uint16) []uint16
-	I2cWrite([]uint16)
+	I2cRead(uint) []byte
+	I2cWrite([]byte)
 }
